Fix #30
